refactor(filemanager): move event message to separate module
