+++ lib/workload/stateless/filemanager/filemanager/src/events/aws/mod.rs
@@ -2,18 +2,17 @@
 //!
 
 use aws_sdk_s3::types::StorageClass as AwsStorageClass;
-use chrono::{DateTime, ParseError, Utc};
+use chrono::{DateTime, Utc};
 use itertools::{izip, Itertools};
-use serde::{Deserialize, Serialize};
+use message::EventMessage;
+use serde::Deserialize;
 use sqlx::postgres::{PgHasArrayType, PgTypeInfo};
 use uuid::Uuid;
 
-use crate::error::Error;
-use crate::error::Error::DeserializeError;
-use crate::error::Result;
 use crate::events::aws::EventType::{Created, Deleted, Other};
 
 pub mod collecter;
+pub mod message;
 
 /// A wrapper around AWS storage types with sqlx support.
 #[derive(Debug, Eq, PartialEq, PartialOrd, Ord, Clone, sqlx::Type)]
@@ -232,7 +231,7 @@ impl From<FlatS3EventMessages> for Events {
 
 /// Flattened AWS S3 events
 #[derive(Debug, Deserialize, Eq, PartialEq, Default)]
-#[serde(try_from = "S3EventMessage")]
+#[serde(try_from = "EventMessage")]
 pub struct FlatS3EventMessages(pub Vec<FlatS3EventMessage>);
 
 impl FlatS3EventMessages {
@@ -490,106 +489,6 @@ impl FlatS3EventMessage {
     }
 }
 
-/// The basic AWS S3 Event.
-#[derive(Debug, Serialize, Deserialize, Eq, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct S3EventMessage {
-    #[serde(rename = "Records")]
-    pub records: Vec<Record>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Eq, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct Record {
-    pub event_time: String,
-    pub event_name: String,
-    pub s3: S3Record,
-}
-
-#[derive(Debug, Serialize, Deserialize, Eq, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct S3Record {
-    pub bucket: BucketRecord,
-    pub object: ObjectRecord,
-}
-
-#[derive(Debug, Serialize, Deserialize, Eq, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct BucketRecord {
-    pub name: String,
-}
-
-#[derive(Debug, Serialize, Deserialize, Eq, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct ObjectRecord {
-    pub key: String,
-    pub size: Option<i32>,
-    pub e_tag: Option<String>,
-    pub version_id: Option<String>,
-    pub sequencer: Option<String>,
-}
-
-impl TryFrom<S3EventMessage> for FlatS3EventMessages {
-    type Error = Error;
-
-    fn try_from(message: S3EventMessage) -> Result<Self> {
-        Ok(FlatS3EventMessages(
-            message
-                .records
-                .into_iter()
-                .map(|record| {
-                    let Record {
-                        event_time,
-                        event_name,
-                        s3,
-                    } = record;
-
-                    let S3Record { bucket, object } = s3;
-
-                    let BucketRecord { name: bucket } = bucket;
-
-                    let ObjectRecord {
-                        key,
-                        size,
-                        e_tag,
-                        version_id,
-                        sequencer,
-                    } = object;
-
-                    let event_time: DateTime<Utc> = event_time
-                        .parse()
-                        .map_err(|err: ParseError| DeserializeError(err.to_string()))?;
-
-                    let event_type = if event_name.contains("ObjectCreated") {
-                        Created
-                    } else if event_name.contains("ObjectRemoved") {
-                        Deleted
-                    } else {
-                        Other
-                    };
-
-                    Ok(FlatS3EventMessage {
-                        s3_object_id: Uuid::new_v4(),
-                        event_time: Some(event_time),
-                        bucket,
-                        key,
-                        size,
-                        e_tag,
-                        sequencer,
-                        version_id,
-                        // Head field are fetched later.
-                        storage_class: None,
-                        last_modified_date: None,
-                        event_type,
-                        number_reordered: 0,
-                        number_duplicate_events: 0,
-                    })
-                })
-                .collect::<Result<Vec<FlatS3EventMessage>>>()?,
-        ))
-    }
-}
-
 impl From<Vec<FlatS3EventMessages>> for FlatS3EventMessages {
     fn from(messages: Vec<FlatS3EventMessages>) -> Self {
         FlatS3EventMessages(messages.into_iter().flat_map(|message| message.0).collect())
@@ -601,9 +500,9 @@ pub(crate) mod tests {
     use chrono::{DateTime, Utc};
     use serde_json::{json, Value};
 
+    use crate::events::aws::message::sqs::SQSEventMessage;
     use crate::events::aws::{
-        EventType, Events, FlatS3EventMessage, FlatS3EventMessages, S3EventMessage,
-        TransposedS3EventMessages,
+        EventType, Events, FlatS3EventMessage, FlatS3EventMessages, TransposedS3EventMessages,
     };
 
     pub(crate) const EXPECTED_SEQUENCER_CREATED_ZERO: &str = "0055AED6DCD90281E3"; // pragma: allowlist secret
@@ -799,7 +698,7 @@ pub(crate) mod tests {
     }
 
     fn expected_flat_events(records: String) -> FlatS3EventMessages {
-        let events: S3EventMessage = serde_json::from_str(&records).unwrap();
+        let events: SQSEventMessage = serde_json::from_str(&records).unwrap();
         events.try_into().unwrap()
     }
 
@@ -815,10 +714,6 @@ pub(crate) mod tests {
     pub(crate) fn expected_events_simple() -> Events {
         expected_events(expected_event_record_simple())
     }
-
-    // pub(crate) fn expected_flat_events_full() -> FlatS3EventMessages {
-    //     expected_flat_events(expected_event_record_full())
-    // }
 
     pub(crate) fn expected_events_full() -> Events {
         expected_events(expected_event_record_full())
